+++ build.gradle
@@ -28,18 +28,14 @@ buildscript {
     }
 }
 
-configurations.compile {
-    exclude group: 'org.springframework.boot', module: 'spring-boot-starter-tomcat'
-}
-
 repositories {
     mavenCentral()
     mavenLocal()
 }
 
 dependencies {
-    compile group: 'com.agonyengine', name: 'agony-forge-core', version: '0.1.0-SNAPSHOT'
-    
+    compile group: 'com.agonyforge', name: 'agony-forge-core', version: '0.1.0-SNAPSHOT'
+
     testCompile group: 'junit', name: 'junit', version: '4.12'
     testCompile group: 'org.mockito', name: 'mockito-core', version: '2.28.2'
 }
